pyproject.toml
[project]
name = "OCB-NASA-APOD"
version = "0.2.0"
description = ""
authors = [{ name = "NodeWee", email = "nodewee@gmail.com" }]
dependencies = ["httpx>=0.23.0", "pillow>=9.2.0", "blurhash-python>=1.1.3"]
requires-python = ">=3.9"
license = { text = "GNU General Public License v3.0" }

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]

[tool.pdm]
